+++ spaced-repetition-back-end/routes/decks/decksRoutes.js
@@ -1,103 +1,35 @@
 const express = require('express');
 const decks = require('./decksModel.js');
+const checkJwt = require('../../jwt');
+const jwtAuthz = require('express-jwt-authz');
 
 const router = express.Router();
+router.use(checkJwt);
 
-// --- FOR TESTING PURPOSES ONLY --- //
+// Should retrieve array of all the user's decks
+// decks should have property cards which is array of all cards
+// omit the user's userid from the response (?)
 router.get('/', (req, res) => {
+  let user_id = req.user.sub
   decks
-    .find()
+    .findByAuthor(user_id)
     .then(decks => {
       res.status(200).json(decks);
-    })
-    .catch(err => res.status(500).json(err));
-});
-
-router.get('/:id', (req, res) => {
-  decks
-    .findById(req.params.id)
-    .then(decks => {
-      res.status(200).json(decks);
-    })
-    .catch(err => res.status(500).json(err));
-});
-// --- END FOR TESTING PURPOSES ONLY --- //
-
-// THIS SHOULD BE AT THE /api/users/:id/decks ENDPOINT
-// This endpoint should also include any matches from the userdeck junction table
-
-router.get('/author/:id', (req, res) => {
-  decks
-    .findByAuthor(req.params.id)
-    .then(decks => {
-      res.status(200).json(decks);
-    })
-    .catch(err => res.status(500).json(err));
-});
-
-router.get('/jct/:id', (req, res) => {
-  decks
-    .findByJct(req.params.id)
-    .then(decks => {
-      res.status(200).json(decks);
-    })
-    .catch(err => res.status(500).json(err));
-});
-
-router.get('/test/:id', (req, res) => {
-  function format(arr) {
-    let deckNames = {};
-    let formattedData = [];
-    let count = 0;
-    for (let i = 0; i < arr.length; i++) {
-      // if deck exists, push just the card to the object's card array
-      if (deckNames[arr[i].name]) {
-        formattedData[deckNames[arr[i].name]].cards.push({
-            "id": arr[i].id,
-            "title": arr[i].title,
-            "question": arr[i].question,
-            "answer": arr[i].answer,
-            "language": arr[i].language
-          })
-      } else {
-        // if deck does not exist, push the deck to formattedData array
-        // add property to deckname objects and assign value of count (for referencing in the array)
-        deckNames[arr[i].name] = count++;
-        formattedData.push({
-          "id": arr[i].deck_id,
-          "name": arr[i].name,
-          "public": arr[i].public,
-          "author": arr[i].author,
-          "user_id": arr[i].user_id,
-          "cards": [{
-            "id": arr[i].id,
-            "title": arr[i].title,
-            "question": arr[i].question,
-            "answer": arr[i].answer,
-            "language": arr[i].language
-          }] 
-        })
-      }
-    }
-    return formattedData;
-  }
-  decks
-    .cardsArrTest(req.params.id)
-    .then(decks => {
-      res.status(200).json(format(decks));
     })
     .catch(err => res.status(500).json(err));
 });
 
 router.post('/', (req, res) => {
-  const user = req.body;
+  const deck = req.body;
+  deck.author = req.user.sub;
 
   decks
-    .add(user)
+    .add(deck)
     .then(ids => {
       res.status(201).json(ids[0]);
     })
     .catch(err => {
+      console.log(err.message);
       res.status(500).json(err);
     });
 });
@@ -132,5 +64,24 @@ router.delete('/:id', (req, res) => {
     })
     .catch(err => res.status(500).json(err));
 });
+
+// ----- Broken after refactor, but planning to fix ------ //
+// router.get('/jct/:id', (req, res) => {
+//   decks
+//     .findByJct(req.params.id)
+//     .then(decks => {
+//       res.status(200).json(decks);
+//     })
+//     .catch(err => res.status(500).json(err));
+// });
+
+// router.get('/test/:id', (req, res) => {
+//   decks
+//     .cardsArrTest(req.params.id)
+//     .then(decks => {
+//       res.status(200).json(format(decks));
+//     })
+//     .catch(err => res.status(500).json(err));
+// });
 
 module.exports = router;
